changed int i and j to long unsigned int to fix compatibility error

+++ 0-bubble_sort.c
@@ -2,7 +2,7 @@
 
 void bubble_sort(int *array, size_t size)
 {
-    int i, j;
+    long unsigned int i, j;
     int temp;
     for (i = 0; i < size - 1; i++)
     {
